More strict types

+++ include/SoloMesh.h
@@ -26,15 +26,15 @@ namespace solo
 
         SL_NONCOPYABLE(Mesh)
 
-        int addBuffer(const VertexBufferLayout& layout, const float* data, int elementCount);
-        void removeBuffer(int index);
+        uint32_t addBuffer(const VertexBufferLayout& layout, const float* data, uint32_t elementCount);
+        void removeBuffer(uint32_t index);
 
-        int addPart(const void* indexData, int indexElementCount);
-        void removePart(int index);
-        int getPartCount() const;
+        uint32_t addPart(const void* indexData, uint32_t indexElementCount);
+        void removePart(uint32_t index);
+        uint32_t getPartCount() const;
 
         void draw(Effect* effect);
-        void drawPart(Effect* effect, uint16_t part);
+        void drawPart(Effect* effect, uint32_t part);
 
         void setPrimitiveType(PrimitiveType type);
         PrimitiveType getPrimitiveType() const;
@@ -53,8 +53,8 @@ namespace solo
         PrimitiveType primitiveType = PrimitiveType::Triangles;
         std::vector<VertexBufferHandle> vertexBuffers;
         std::vector<IndexBufferHandle> indexBuffers;
-        std::vector<int> vertexCounts;
-        int minVertexCount = 0;
+        std::vector<uint32_t> vertexCounts;
+        uint32_t minVertexCount = 0;
         VertexObjectHandle vertexObjectHandle;
         VertexObjectHandle effectBindingVertexObjectHandle;
     };
@@ -69,8 +69,8 @@ namespace solo
         return primitiveType;
     }
 
-    inline int Mesh::getPartCount() const
+    inline uint32_t Mesh::getPartCount() const
     {
-        return static_cast<int>(indexBuffers.size());
+        return static_cast<uint32_t>(indexBuffers.size());
     }
 }
